# docs/api/tcheck-corporate.openapi.yaml
openapi: 3.0.3
info:
  title: TallyCheck Corporate API
  description: |
    Corporate attendance, SafeChild pickup, BLE beacons and tenant settings.

    This is one of several TallyCheck service contracts. It does **not** cover the
    academic domain (courses, cohorts, enrolment) — that lives in `Tcheck-Backend`
    and is described by `tcheck-academic.openapi.yaml`. Visitor management lives in
    `vcheck.openapi.yaml`. The three are separate documents on purpose: they run on
    different hosts, use different auth, and have colliding paths (`/attendance`
    means corporate clock-in here and class attendance in the academic service).

    ## Multi-tenancy
    Every request resolves to exactly one tenant schema before any handler runs.
    See `shared/tenancy.yaml`. Callers holding a token send nothing extra — the
    token carries the tenant.

    ## Roles
    Roles are enumerated once, in `shared/primitives.yaml#/components/schemas/RoleKey`,
    and must match `libs/auth/src/roles.ts` and `backend/helpers/auth_helper.py`.
    Some tenants still carry the legacy names `hr` and `manager`; the backend
    guards accept both.

    ## Authentication
    All endpoints except `GET /health` and
    `GET /auth/organization-by-subdomain/{subdomain}` require a Bearer JWT.

    ## Response shape
    Successful responses are enveloped as `{ "message": ..., "data": ... }`, with
    `meta` added on paginated collections. See `shared/envelope.yaml`.
  version: 2.1.0
servers:
...
security:
  - Auth0: []

# ==========================================================
# REUSABLE STRUCTURES
# ==========================================================
components:
  securitySchemes:
    Auth0:
      type: openIdConnect
      # Must match the tenant configured in apps/intranet/src/app/app.config.ts.
      # This previously named a different tenant than the app actually used.
      openIdConnectUrl: https://dev-x506paw8afbw6jgu.us.auth0.com/.well-known/openid-configuration

  # ==========================================================
  # REUSABLE SCHEMAS
  # ==========================================================
  schemas:
...
    ErrorResponse:
      $ref: './shared/errors.yaml#/components/schemas/ErrorResponse'
    PaginationMeta:
      $ref: './shared/envelope.yaml#/components/schemas/Meta'
...
    AttendanceRecord:
      type: object
      properties:
        id:
          type: integer
          example: 101
        employee_id:
          type: integer
          example: 12
        clock_in:
          type: string
          format: date-time
          example: "2026-06-10T08:02:11Z"
        clock_out:
          type: string
          format: date-time
          nullable: true
          example: "2026-06-10T17:05:44Z"
        work_date:
          type: string
          format: date
          example: "2026-06-10"
        worked_hours:
          type: number
          format: float
          nullable: true
          example: 9.06
        source:
          type: string
          example: "web"
        status:
          type: string
          enum: [open, closed, corrected]
          example: "closed"
        notes:
          type: string
          nullable: true
          example: null
        edited_by:
          type: integer
          nullable: true
          description: "Employee ID of the HR/Admin who last edited this record"
    ClockInResponse:
      type: object
      properties:
        record:
          $ref: "#/components/schemas/AttendanceRecord"
        message:
          type: string
          example: "Clocked in at 08:02 UTC"
    ClockOutResponse:
      type: object
      properties:
        record:
          $ref: "#/components/schemas/AttendanceRecord"
        message:
          type: string
          example: "Clocked out. Worked 9.06 hours."
    AttendanceListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AttendanceRecord"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
...
    AttendanceManualRequest:
      type: object
      required: [employee_id, work_date, clock_in, clock_out, reason]
      description: Manually create and attendance tackle on behalf of an employee
      properties:
        employee_id:
          type: integer
          example: 12
        work_date:
          type: string
          format: date
          example: "2026-06-09"
        clock_in:
          type: string
          format: date-time
          example: "2026-06-09T08:00:00Z"
        clock_out:
          type: string
          format: date-time
          nullable: true
          example: "2026-06-09T17:00:00Z"
        reason:
          type: string
          example: "Employee worked on-site without system access."

    # ==========================================================
    # EMPLOYEE
    # ==========================================================
...
    ReportEmployeeAttendance:
      type: object
      properties:
        employee_id:
          type: integer
          example: 12
        employee_name:
          type: string
          example: "Nyambane Mogaka"
        department:
          type: string
          example: "software_development"
        total_days_present:
          type: integer
          example: 18
        total_days_absent:
          type: integer
          example: 2
        total_hours_worked:
          type: number
          format: float
          example: 162.5
        avg_clock_in:
          type: string
          example: "08:07"
        avg_clock_out:
          type: string
          example: "17:23"
        late_days:
          type: integer
          example: 3
    ReportDepartmentAttendance:
      type: object
      properties:
        department_id:
          type: integer
          example: 2
        department_name:
          type: string
          example: "Software Development"
        date_from:
          type: string
          format: date
        date_to:
          type: string
          format: date
        total_employees:
          type: integer
          example: 4
        avg_attendance_rate:
          type: number
          format: float
          description: "Percentage 0–100"
          example: 94.5
        total_hours_worked:
          type: number
          format: float
          example: 640.5
        employees:
          type: array
          items:
            $ref: "#/components/schemas/ReportEmployeeAttendance"
...
    TimelineEntry:
      type: object
      properties:
        employee_id:
          $ref: './shared/primitives.yaml#/components/schemas/Uuid'
        first_name:
          type: string
        last_name:
          type: string
        initials:
          type: string
          example: JM
        department:
          type: string
          description: Falls back to "Unassigned".
          example: software_development
        records:
          type: array
          items:
            type: object
            properties:
              clock_in:
                type: string
                format: date-time
                nullable: true
              clock_out:
                type: string
                format: date-time
                nullable: true
              status:
                type: string
              notes:
                type: string
                nullable: true
    TrendPoint:
      type: object
      properties:
        date:
          $ref: './shared/primitives.yaml#/components/schemas/IsoDate'
        total_hours:
          type: number
          format: float
          description: Rounded to 2dp. Zero for days with no completed sessions.
          example: 76.25

    # ==========================================================
    # BEACONS
    # ==========================================================
...
  responses:
    Unauthorized: # Can be referenced as '#/components/responses/unauthorized'
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Conflict:
      description: Business rule conflict (e.g. already clocked in)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnprocessableEntity:
      description: Validation error e.g. missing required field
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  # ==========================================================
  # REUSABLE PARAMETERS
  # ==========================================================
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
    DateFromParam:
      name: date_from
      in: query
      schema:
        type: string
        format: date
      example: "2026-06-01"
    DateToParam:
      name: date_to
      in: query
      schema:
        type: string
        format: date
      example: "2026-06-30"
    EmployeeIdParam:
      name: employee_id
      in: query
      schema:
        type: integer
    DepartmentIdParam:
      name: department_id
      in: query
      schema:
        type: integer
...
paths:
...
  /attendance/clock-in:
    post:
      tags: [Attendance]
      summary: Clock in (Employee)
      operationId: clock_in
      description: |
        Records a server-side clock-in timestamp for the authenticated employee.
        Returns `409` if the employee already has an open (un-closed) record for today.
      responses:
        "201":
          description: Clocked in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClockInResponse"
        "409":
          description: Already clocked in, an open record exists
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /attendance/clock-out:
    post:
      tags: [Attendance]
      summary: Clock out (Employee)
      operationId: clock_out
      description: |
        Closes the current open attendance record. Returns `409` if no open record
        exists (i.e. employee has not clocked in).
      responses:
        "200":
          description: Clocked out successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClockOutResponse"
        "409":
          description: No open record exists
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /attendance:
    get:
      tags: [HR - Attendance]
      summary: Query attendance records
      operationId: list_attendance
      description: |
        Managers - retrieve attendance records for their department
        HR - retrieve attendance records for all employees
      parameters:
        - $ref: "#/components/parameters/EmployeeIdParam"
        - $ref: "#/components/parameters/DepartmentIdParam"
        - $ref: "#/components/parameters/DateFromParam"
        - $ref: "#/components/parameters/DateToParam"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PerPageParam"
        - name: status
          in: query
          schema:
            type: string
            enum: [open, closed, corrected]
      responses:
        "200":
          description: Filtered, paginated attendance records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [HR - Attendance]
      summary: Manually create an attendance record (HR only)
      operationId: create_attendance_manual
      description: |
        Backfill or manually enter an attendance record on behalf of an employee.
        Writes an audit log entry with the mandatory `reason` field.
        Sets that attendance record's status to 'corrected.'
        Only done by HR & Admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceManualRequest"
      responses:
        "201":
          description: Manual record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: A record for that employee/date already exists
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
...
  /reports/attendance:
    get:
      tags: [Reports]
      summary: Attendance summary report (HR only)
      operationId: report_attendance
      description: |
        Returns per-employee attendance summaries for the given date range.
        Add `?format=csv` to download as a CSV file instead of JSON.
        Only done by HR & Admin
      parameters:
        - $ref: "#/components/parameters/DateFromParam"
        - $ref: "#/components/parameters/DateToParam"
        - $ref: "#/components/parameters/DepartmentIdParam"
        - $ref: "#/components/parameters/EmployeeIdParam"
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv]
            default: json
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PerPageParam"
      responses:
        "200":
          description: Attendance summary (JSON or CSV)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ReportEmployeeAttendance"
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"
            text/csv:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /reports/attendance/departments:
    get:
      tags: [Reports]
      summary: Get attendance report for all departments (HR only)
      operationId: report_attendance_all_departments
      parameters:
        - $ref: "#/components/parameters/DateFromParam"
        - $ref: "#/components/parameters/DateToParam"
      responses:
        "200":
          description: Departments attendance records found
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ReportDepartmentAttendance"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /reports/attendance/departments/{id}:
    get:
      tags: [Reports]
      summary: Get attendance report for one department (HR only)
      operationId: report_attendance_one_department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
        - $ref: "#/components/parameters/DateFromParam"
        - $ref: "#/components/parameters/DateToParam"
      responses:
        "200":
          description: Department attendance record found
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ReportDepartmentAttendance"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /reports/dashboard:
    get:
      tags: [Reports]
      summary: HR dashboard KPIs
      operationId: report_dashboard
      description: |
        Returns today's real-time headcount metrics.
        Only done by HR & Admin
      responses:
        "200":
          description: Dashboard KPIs
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_headcount:
                    type: integer
                    example: 47
                  present_today:
                    type: integer
                    example: 39
                  absent_today:
                    type: integer
                    example: 8
                  currently_clocked_in:
                    type: integer
                    example: 34
                  attendance_rate_today:
                    type: number
                    format: float
                    description: "Percentage 0–100"
                    example: 82.98
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ==========================================================
  # AUTH — tenant resolution, profile sync, approval queue
  # ==========================================================
...
  /reports/attendance/timeline:
    get:
      tags: [Reports]
      summary: One day of attendance across all employees
      description: >
        Employees with no department are excluded, which in practice filters out
        platform super admins.
      parameters:
        - name: date
          in: query
          required: false
          description: Defaults to today.
          schema:
            $ref: './shared/primitives.yaml#/components/schemas/IsoDate'
      responses:
        "200":
          description: Per-employee timeline for the target date.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: './shared/envelope.yaml#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/TimelineEntry"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /reports/attendance/trends:
    get:
      tags: [Reports]
      summary: Total hours per day over a range
      description: >
        Every day in the range is present, including days with no records, which
        report `total_hours: 0` — so the series can be charted without gap filling.
      parameters:
        - name: date_from
          in: query
          schema:
            $ref: './shared/primitives.yaml#/components/schemas/IsoDate'
        - name: date_to
          in: query
          schema:
            $ref: './shared/primitives.yaml#/components/schemas/IsoDate'
        - name: department_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Daily totals, ascending.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: './shared/envelope.yaml#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/TrendPoint"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ==========================================================
  # BEACONS
  # ==========================================================
...