# libs/api-client/src/generated/tcheck/model/trendPoint.ts
/**
 * Generated by orval from TallyCheck Corporate API v2.1.0.
 * DO NOT EDIT — run `nx run api-client:generate` instead.
 */
import type { IsoDate } from "./isoDate";

export interface TrendPoint {
  date?: IsoDate;
  /** Rounded to 2dp. Zero for days with no completed sessions. */
  total_hours?: number;
}